AI-ML-DEEP-LEARNING-ENGINE/requirements.txt
# 🌟 AI/ML Deep Learning Engine - Python Dependencies

# ===================================
# CORE AI/ML FRAMEWORKS
# ===================================
torch>=2.1.0
tensorflow>=2.15.0
transformers>=4.36.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.23

# ===================================
# LLM PROVIDERS
# ===================================
openai>=1.12.0
anthropic>=0.18.0
google-generativeai>=0.3.2
cohere>=4.47
together>=0.2.0

# ===================================
# VECTOR DATABASES
# ===================================
pinecone-client>=3.0.0
weaviate-client>=4.4.0
qdrant-client>=1.7.0
pymilvus>=2.3.5
chromadb>=0.4.22
faiss-cpu>=1.7.4

# ===================================
# EMBEDDINGS
# ===================================
sentence-transformers>=2.3.0
tiktoken>=0.5.2
openai-embeddings>=0.1.0

# ===================================
# DATA PROCESSING
# ===================================
numpy>=1.26.0
pandas>=2.1.4
scipy>=1.11.4
scikit-learn>=1.4.0

# ===================================
# NLP & TEXT PROCESSING
# ===================================
spacy>=3.7.0
nltk>=3.8.1
textblob>=0.17.1
beautifulsoup4>=4.12.3
unstructured>=0.12.0

# ===================================
# COMPUTER VISION
# ===================================
opencv-python>=4.9.0
pillow>=10.2.0
torchvision>=0.16.0

# ===================================
# GRAPH & KNOWLEDGE BASE
# ===================================
neo4j>=5.16.0
networkx>=3.2.1
rdflib>=7.0.0

# ===================================
# API FRAMEWORK
# ===================================
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.26.0

# ===================================
# ASYNC & CONCURRENCY
# ===================================
asyncio>=3.4.3
aiohttp>=3.9.1
celery>=5.3.4

# ===================================
# DATABASE DRIVERS
# ===================================
psycopg2-binary>=2.9.9
pymongo>=4.6.1
redis>=5.0.1
elasticsearch>=8.11.1

# ===================================
# MESSAGE QUEUES
# ===================================
kafka-python>=2.0.2
pika>=1.3.2

# ===================================
# MONITORING & LOGGING
# ===================================
prometheus-client>=0.19.0
python-json-logger>=2.0.7
structlog>=24.1.0

# ===================================
# UTILITIES
# ===================================
python-dotenv>=1.0.1
pyyaml>=6.0.1
requests>=2.31.0
tenacity>=8.2.3
tqdm>=4.66.1

# ===================================
# TESTING
# ===================================
pytest>=8.0.0
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0

# ===================================
# AUTO ML
# ===================================
optuna>=3.5.0
ray[tune]>=2.9.0
hyperopt>=0.2.7

# ===================================
# REINFORCEMENT LEARNING
# ===================================
gymnasium>=0.29.1
stable-baselines3>=2.2.1

# ===================================
# DEPLOYMENT
# ===================================
gunicorn>=21.2.0
python-multipart>=0.0.6